Shader code for new approach to defend nodes
Rust code to set up and call the shader is still TBD.

Previously, the shader defend_combine.wgsl would create all combinations
of the energies under each successor at once and then minimize those.
This does not scale well, because the number of all combinations is
equal to the product of the number of energies at each successor. This
quickly gets out of hand.

The new approach is based on Proposition 11 in the paper "Multi-weighted
reachability games" by T. Brihaye and A. Goeminne, 2023.
Instead of creating all combinations at once, we now iterate over the
successors and only create combinations with the energies of the next
successor. The resulting energies are minimized at each iteration which
should keep the runtime down even on larger inputs.

The downside is that it is less parallelized and requires some
workarounds to work in a shader, but the reduced number of energies we
have to deal with at once should far outweigh any such losses.