Ensure uniformity of control barriers in shaders
Control barriers (such as `workgroupBarrier()` and `storageBarrier()`)
must be executed by all threads in a workgroup in lock-step.
To comply with that, this commit adjusts the shaders so that barriers
never occur e.g. behind an if-statement that excludes some threads of a
workgroup.

This solves GPU runtime errors that were observed on Intel,
but not on AMD hardware, both running the mesa driver stack on Linux.